requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
dnspython==1.16.0
get==2019.4.13
idna==2.8
IP2Location==8.4.1
netifaces==0.10.9
Pillow==7.0.0
post==2019.4.13
public==2019.4.13
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
requests-file==1.4.3
six==1.14.0
socket.py==0.1.0
sockets==1.0.0
soupsieve==1.9.5
tldextract==2.2.2
urllib3==1.25.7
urlopen==1.0.0
wincertstore==0.2
